Add files via upload
Problem Statement

BoomBikes aspires to understand the demand for shared bikes among the people so as to strategise business plan after this ongoing quarantine situation ends across the nation due to Covid-19.

They want to understand the factors on which the demand for these shared bikes depends. Specifically, they want to understand the factors affecting the demand for these shared bikes in the American market. The company wants to know:

Which variables are significant in predicting the demand for shared bikes.
How well those variables describe the bike demands